Indefinitely stopping development on this project.
Version: ~
Author(s): Anthony Pizzimenti
Date: November 20, 2017

============================= NOTES ===================================
This is the current and, in all likelihood, last iteration of this
project. I am going to halt development indefinitely, mostly because the
purpose of this project was for a specific game and framework. Someone
else using this or taking over development isn't likely, but I'll still
be working on trying to get this merged with its parent project.

Anthony